Added Inventory.
