--- src/main/java/javaclasses/library/Inventory/BookMarkedAsAvailable.java
@@ -1,22 +1,0 @@
-package javaclasses.library.Inventory;
-
-import javaclasses.library.tinytype.InventoryId;
-import javaclasses.library.tinytype.InventoryItemId;
-
-public class BookMarkedAsAvailable {
-    private final InventoryId inventoryId;
-    private final InventoryItemId inventoryItemId;
-
-    public BookMarkedAsAvailable(InventoryId inventoryId, InventoryItemId inventoryItemId) {
-        this.inventoryId = inventoryId;
-        this.inventoryItemId = inventoryItemId;
-    }
-
-    public InventoryId getInventoryId() {
-        return inventoryId;
-    }
-
-    public InventoryItemId getInventoryItemId() {
-        return inventoryItemId;
-    }
-}
