+++ src/main/java/javaclasses/library/inventory/BookMarkedAsLost.java
@@ -1,18 +1,21 @@
-package javaclasses.library.Inventory;
+package javaclasses.library.inventory;
 
 import javaclasses.library.tinytype.InventoryId;
 import javaclasses.library.tinytype.InventoryItemId;
 import javaclasses.library.tinytype.UserId;
 
-public class BookReadyToPickup {
+public class BookMarkedAsLost {
     private final InventoryId inventoryId;
     private final InventoryItemId inventoryItemId;
     private final UserId userId;
-
-    public BookReadyToPickup(InventoryId inventoryId, InventoryItemId inventoryItemId, UserId userId) {
+    public BookMarkedAsLost(InventoryId inventoryId, InventoryItemId inventoryItemId, UserId userId) {
         this.inventoryId = inventoryId;
         this.inventoryItemId = inventoryItemId;
         this.userId = userId;
+    }
+
+    public UserId getUserId() {
+        return userId;
     }
 
     public InventoryId getInventoryId() {
@@ -21,9 +24,5 @@ public class BookReadyToPickup {
 
     public InventoryItemId getInventoryItemId() {
         return inventoryItemId;
-    }
-
-    public UserId getUserId() {
-        return userId;
     }
 }
