+++ src/main/java/javaclasses/library/inventory/BookReadyToPickup.java
@@ -1,21 +1,18 @@
-package javaclasses.library.Inventory;
+package javaclasses.library.inventory;
 
 import javaclasses.library.tinytype.InventoryId;
 import javaclasses.library.tinytype.InventoryItemId;
 import javaclasses.library.tinytype.UserId;
 
-public class BookMarkedAsLost {
+public class BookReadyToPickup {
     private final InventoryId inventoryId;
     private final InventoryItemId inventoryItemId;
     private final UserId userId;
-    public BookMarkedAsLost(InventoryId inventoryId, InventoryItemId inventoryItemId, UserId userId) {
+
+    public BookReadyToPickup(InventoryId inventoryId, InventoryItemId inventoryItemId, UserId userId) {
         this.inventoryId = inventoryId;
         this.inventoryItemId = inventoryItemId;
         this.userId = userId;
-    }
-
-    public UserId getUserId() {
-        return userId;
     }
 
     public InventoryId getInventoryId() {
@@ -24,5 +21,9 @@ public class BookMarkedAsLost {
 
     public InventoryItemId getInventoryItemId() {
         return inventoryItemId;
+    }
+
+    public UserId getUserId() {
+        return userId;
     }
 }
