+++ src/main/java/javaclasses/library/inventory/InventoryAppended.java
@@ -1,23 +1,22 @@
-package javaclasses.library.Inventory;
+package javaclasses.library.inventory;
 
 import javaclasses.library.tinytype.InventoryItemId;
 import javaclasses.library.tinytype.InventoryId;
 
-public class InventoryDecreased {
+public class InventoryAppended {
     private final InventoryItemId inventoryItemId;
     private final InventoryId inventoryId;
 
-    public InventoryDecreased(InventoryItemId inventoryItemId, InventoryId inventoryId) {
+    public InventoryAppended(InventoryItemId inventoryItemId, InventoryId inventoryId) {
         this.inventoryItemId = inventoryItemId;
         this.inventoryId = inventoryId;
-
-    }
-
-    public InventoryItemId getInventoryItemId() {
-        return inventoryItemId;
     }
 
     public InventoryId getInventoryId() {
         return inventoryId;
+    }
+
+    public InventoryItemId getInventoryItemId() {
+        return inventoryItemId;
     }
 }
