+++ src/main/java/javaclasses/library/inventory/ReservationRemoved.java
@@ -1,13 +1,13 @@
-package javaclasses.library.Inventory;
+package javaclasses.library.inventory;
 
 import javaclasses.library.tinytype.InventoryId;
 import javaclasses.library.tinytype.UserId;
 
-public class ReservationCreated {
+public class ReservationRemoved {
     private final InventoryId inventoryId;
     private final UserId userId;
 
-    public ReservationCreated(InventoryId inventoryId, UserId userId) {
+    public ReservationRemoved(InventoryId inventoryId, UserId userId) {
         this.inventoryId = inventoryId;
         this.userId = userId;
     }
